Fixed Tests
Helpers was defined as a module and a file
